fix(coach): move InterviewFormat into Config and drop Bundle.module
Bundle.module's generated accessor only checks Bundle.main.bundleURL (the
.app root, not Contents/Resources) or a compile-time absolute path, and
fatalErrors when neither exists — a real install has neither, so selecting
a format with content would crash. Replace it with a manual search mirroring
SileroVoiceActivityDetector.bundledModelURL(): Bundle.main.resourceURL for a
shipped app, Bundle.main.bundleURL for swift build/run, and a #filePath-relative
read of the checked-out source for swift test (where Bundle.main resolves to
the unrelated swiftpm-testing-helper).

FileManager-based file I/O also doesn't belong in Sources/JarvisCore/Coach,
which scripts/check-coaching-kernel.sh enforces as OS-reach-through-free.
Move InterviewFormat, CoachingSkill, and their tests to Config/, which the
kernel hands a frozen snapshot rather than owning; CoachDriver and
CoachAttemptRunner only ever see a pre-resolved String, never the type.

Also correct wiki/architecture.md's interview-format paragraph, which
listed Coding/Behavioral as picker options despite their empty addenda
being filtered out, and overclaimed that dropping in a new <format>.md
alone creates a picker entry — it only does for an existing case;  a new
format still needs a new InterviewFormat case and display name first.
